requirements.txt
requests
speedtest-cli
addict